requirements.txt
avro==1.11.1
dpkt==1.9.7.2
lxml==4.9.1
numpy==1.23.1
fastavro==1.5.4
six>=1.12.0
google-cloud-pubsub==2.13.4
google-cloud-bigquery==3.3.0